+++ tests/test_optimizer.py
@@ -27,7 +27,7 @@ class TestOptreex:
         optax_params: tx.Linear
         optax_updates, opt_state = optax_optim.update(grads, opt_state, linear)
         optax_params = optax.apply_updates(optax_updates, linear)
-        treex_params = optimizer.apply_updates(grads, linear)
+        treex_params = optimizer.update(grads, linear)
 
         assert all(
             np.allclose(a, b)
@@ -59,7 +59,7 @@ class TestOptreex:
 
         optax_updates: tx.Linear
         optax_updates, opt_state = optax_optim.update(grads, opt_state, linear)
-        treex_updates = optimizer.apply_updates(grads, linear, return_updates=True)
+        treex_updates = optimizer.update(grads, linear, return_updates=True)
 
         assert all(
             np.allclose(a, b)
