+++ treex/types.py
@@ -1,4 +1,3 @@
-import dataclasses
 import typing as tp
 from dataclasses import dataclass
